Fix issue with rolling restart on Ubuntu 16.4.
Since Ubuntu 15.4, Ubuntu uses System D instead of Upstart.
Turns out there's different semantics: Upstart used to pass the
arguments to the script (in this case 'restart') but System D does a
'stop' followed by an immediate 'start' when invoking 'restart'.
As a result, Unicorn's master process is signalled with QUIT and before
it has time to stop, System D tries to start a new process using the
script but sees the process (still) running and thinks it shouldn't do
anything. As a result the site goes down.